Unidirectional registration
