+++ modules/sdk/src/main/scala/org/tessellation/sdk/infrastructure/consensus/ConsensusManager.scala
@@ -2,7 +2,7 @@ package org.tessellation.sdk.infrastructure.consensus
 
 import cats.data.Ior.{Both, Right}
 import cats.effect._
-import cats.effect.std.{Queue, Supervisor}
+import cats.effect.std.Supervisor
 import cats.kernel.Next
 import cats.syntax.applicativeError._
 import cats.syntax.contravariantSemigroupal._
@@ -21,17 +21,17 @@ import scala.reflect.runtime.universe.TypeTag
 import org.tessellation.ext.cats.syntax.next._
 import org.tessellation.schema.node.NodeState
 import org.tessellation.schema.node.NodeState.{Observing, Ready}
+import org.tessellation.schema.peer.Peer
 import org.tessellation.schema.peer.Peer.toP2PContext
-import org.tessellation.schema.peer.{Peer, PeerId}
 import org.tessellation.sdk.domain.cluster.storage.ClusterStorage
 import org.tessellation.sdk.domain.gossip.Gossip
 import org.tessellation.sdk.domain.node.NodeStorage
+import org.tessellation.sdk.infrastructure.consensus.registration.Deregistration
 import org.tessellation.sdk.infrastructure.consensus.trigger.{ConsensusTrigger, EventTrigger, TimeTrigger}
 import org.tessellation.sdk.infrastructure.metrics.Metrics
 import org.tessellation.security.signature.Signed
 
 import eu.timepit.refined.auto._
-import fs2.Stream
 import io.circe.{Decoder, Encoder}
 import org.typelevel.log4cats.slf4j.Slf4jLogger
 
@@ -54,17 +54,15 @@ object ConsensusManager {
     nodeStorage: NodeStorage[F],
     clusterStorage: ClusterStorage[F],
     consensusClient: ConsensusClient[F, Key],
-    gossip: Gossip[F],
-    selfId: PeerId
-  )(implicit S: Supervisor[F]): F[ConsensusManager[F, Key, Artifact]] = Queue.unbounded[F, Peer].flatMap { peersForRegistrationExchange =>
+    gossip: Gossip[F]
+  )(implicit S: Supervisor[F]): F[ConsensusManager[F, Key, Artifact]] = {
     val logger = Slf4jLogger.getLoggerFromClass[F](ConsensusManager.getClass)
 
-    def exchangeRegistration(peer: Peer): F[Unit] = {
-      val exchange = for {
-        exchangeRequest <- consensusStorage.getOwnRegistration.map(RegistrationExchangeRequest(_))
-        exchangeResponse <- consensusClient.exchangeRegistration(exchangeRequest).run(peer)
-        maybeResult <- exchangeResponse.maybeKey.traverse(consensusStorage.registerPeer(peer.id, _))
-        _ <- (exchangeResponse.maybeKey, maybeResult).traverseN {
+    def getRegistration(peer: Peer): F[Unit] =
+      for {
+        registrationResponse <- consensusClient.getRegistration.run(peer)
+        maybeResult <- registrationResponse.maybeKey.traverse(consensusStorage.registerPeer(peer.id, _))
+        _ <- (registrationResponse.maybeKey, maybeResult).traverseN {
           case (key, result) =>
             if (result)
               logger.info(s"Peer ${peer.id.show} registered at ${key.show}")
@@ -72,19 +70,6 @@ object ConsensusManager {
               logger.warn(s"Peer ${peer.id.show} cannot be registered at ${key.show}")
         }
       } yield ()
-
-      exchange
-        .handleErrorWith(err => logger.error(err)(s"Error exchanging registration with peer ${peer.show}"))
-    }
-
-    def startRegistrationExchange: F[Unit] =
-      S.supervise {
-        Stream
-          .fromQueueUnterminated(peersForRegistrationExchange)
-          .evalMap(exchangeRegistration)
-          .compile
-          .drain
-      }.void
 
     val manager = new ConsensusManager[F, Key, Artifact] {
 
@@ -94,7 +79,6 @@ object ConsensusManager {
           val facilitationKey = lastKey.nextN(2L)
 
           consensusStorage.setOwnRegistration(facilitationKey) >>
-            startRegistrationExchange >>
             consensusStorage.setLastKey(lastKey) >>
             consensusStorage
               .getResources(observationKey)
@@ -118,7 +102,6 @@ object ConsensusManager {
       def startFacilitatingAfter(lastKey: Key, lastArtifact: Signed[Artifact]): F[Unit] =
         consensusStorage.setLastKeyAndArtifact(lastKey, lastArtifact) >>
           consensusStorage.setOwnRegistration(lastKey.next) >>
-          startRegistrationExchange >>
           scheduleFacility
 
       private def scheduleFacility: F[Unit] =
@@ -239,8 +222,10 @@ object ConsensusManager {
             none[Peer]
         }
           .filter(_.isResponsive)
-          .filter(selfId < _.id)
-          .enqueueUnterminated(peersForRegistrationExchange)
+          .parEvalMapUnbounded { peer =>
+            getRegistration(peer)
+              .handleErrorWith(err => logger.error(err)(s"Error exchanging registration with peer ${peer.show}"))
+          }
           .compile
           .drain
       ).as(manager)
