--- modules/sdk/src/main/scala/org/tessellation/sdk/infrastructure/consensus/PeerRegistration.scala
@@ -1,6 +1,0 @@
-package org.tessellation.sdk.infrastructure.consensus
-
-sealed trait PeerRegistration[Key]
-
-case class Registered[Key](at: Key) extends PeerRegistration[Key]
-case class Deregistered[Key](at: Key) extends PeerRegistration[Key]
